refactor: change name from equatable to extends
