+++ test_project/test/extends/getter_test.dart
@@ -1,4 +1,4 @@
-import 'package:example/equatable/getter.dart';
+import 'package:example/extends/getter.dart';
 import 'package:test/test.dart';
 
 void main() {
